watchman 4.7.0
We've just completed rolling this out internally at FB so we're about ready to release to homebrew.  This PR is to check that we pass the brew CI, then I will tag as `v4.7.0` without the `-rc1` suffix; that will mean changing the URL to match once we've proven this.